[TASK] Refactor package to be decoupled from TYPO3.Party Package
Removed hard dependency on TYPO3.Party by creating a
UserContextServiceInterface, a service implementing this interface
can add additional information to the context data given to the
sentry client.

This refactoring is needed to stay compatible with the current
Flow development.

There is no BC break as the current configuration has the same
behaviour like the last versions